requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2
Pillow==10.0.1
gunicorn==21.2.0
